Use http instead of https for bitcoincharts
This may fix the NaNBTC (and possibly other related issues) issue that people have on OSX. I'm not able to replicate this but if the reports of this problem vanish after including this, then it could be considered fixed. 

bitcoinaverage, bitpay, and blockchain.info all force https but bitcoincharts allows http.

Pro: newbies on OSX won't have to go through the fixing steps detailed here https://github.com/OpenBazaar/OpenBazaar-Client/issues/1247#issuecomment-213123969

Con: there is less redundancy for people with this issue.